remove panel tab UI
